Eighth Push; Development 7 am
Created a file called ScrollToTop that makes sure you stay in screen and at the top of the screen when reloading. Added a left and right barbell for the Landing Page title, and added a back button in the chatbot page to go back to landing page. Some minor tweaks were made with padding and spacing throughout.